페이지 기능 고도화
- 좋아요 버튼 클릭시 비동기 처리 기능 구현
- 회원 가입시 아이디 중복 체크 비동기 처리 구현

+++ src/shop/gigabox/dao/RDao.java
@@ -9,7 +9,7 @@ public interface RDao {
 	public List<RVO> selectReviewListByMvIdx(int mv_idx);
 	public RVO selectReviewByIdx(int r_idx);
 	public int selectCountReviewByIdx(RVO rvo);
-	public int selectAvgReviewScoreByIdx(int mv_idx);
+	public double selectAvgReviewScoreByIdx(int mv_idx);
 	public int insertReview(RVO rvo);
 	public int updateReview(RVO rvo);
 	public int deleteReview(int r_idx);
