+++ src/shop/gigabox/service/RService.java
@@ -9,7 +9,7 @@ public interface RService {
 	public List<RVO> getReviewListByMovie(int mv_idx);
 	public RVO getReviewByIdx(int r_idx);
 	public int checkReview(RVO rvo);
-	public int getReviewScoreAvg(int mv_idx);
+	public double getReviewScoreAvg(int mv_idx);
 	public int insertReview(RVO rvo);
 	public int updateReview(RVO rvo);
 	public int deleteReview(int r_idx);
